ENH: plot time-frequency representations on sensor layout - 2 user functions for power / phase lock - private function that does the actual job - example
For the time being the user functions may seem redundant. However we might want to add autmatic text plotting for more convenience where having two seperate functions make sense.
The the colorbar / borders / scaling issue now is resolved and to me the plot looks fine. However there might be a more elegant way of doing it (scaling the layout and the color scale).
Anyways, If you are fine with the functionality / API, I would include it as is for the forthcoming release. We can later optimize the implementation.