WIP mne_kit2fiff: from KIT SQD files to fif files
Here is the beginning of a python implementation of mne_kit2fiff. It is
very much a WIP. Major points to be worked on are:
-channel coil transformation
-channel unit vectors of coord_frame.
-fix dev_head_t
-creating a synthetic trigger channel

FIX: dev_head_t, RM: excess classes, unneccesary methods

Fixed the error in dev_head_t to give proper transformation matrix.
Reduced the number of classes to three main ones:
-sqd_params, RawKIT, coreg.
Added stringdocs to all methods and classes

ENH: layout

Enhances some of the layout. Update to be done for  docstrings to match
style. PEP8 in progress. :)

ADD: test files for troubleshooting and validation

These are all the required files for creating a raw object.
-sns.txt is the sensor locations for KIT-NYU system.
-test.elp is the digitizer laser points from Polhemus.
-test.hsp is the digitizer head shape from Polhemus.
-test.sqd is the datafile acquired from the KIT-NYU system.
-test_Ykgw.mat is a mat file containing the data extracted using the
Yokogawa matlab toolkit. This is to check that the proper data is being
extracted.
-test_marker.txt is the marker coil position within the MEG.

ENH: pep8 correction

STY: update script

Update mne/bin/mne_kit2fiff.py

Fix error.