ENH: add compability for KIT-NYUAD
ENH: addresses the comments

coreg is now a module of functions.
kit has been condensed and stripped of methods; only relevant Raw
atrributes are added to keep consistency.

Problem to be resolved:
there is a problem with deep copy. Currently there is a hack in place
to deal with it.

ENH: scipy norm instead of numpy FIX: bunch problem