WIP ENH: Resume dataset downloads with NILS code
Copy code from NILS to enable resuming downloads of sample datasets. Use the
NILS code to perform the downloading instead of the previous code.

This is just the first step in implementing resuming dataset downloads. Next
the borrowed code must be integrated with the previously used code. For
example, the progress bar isn't working now.