ENH: Don't create config dir unless necessary
When reading configuration variables which can be set via environment
variables or a config file stored in ~/.mne, don't attempt to create the
config directory or file.  The ~ directory may not be writeable if the
application is run as a daemon (e.g. via uWSGI).  Any attempts to write
to the filesystem should only occur when `set_config` is called.