FIX Compare time boundaries with tolerance in crop
When cropping we want to include both tmin and tmax time points. However, tmin
and tmax might be almost equal, within float accuracy, to one of the stc.times
points. In such a case using np.isclose will now guarantee that both the tmin
and tmax points get included if they're equal (within tolerance) to any of the
time points in stc.times.

This happens for example when iterating through time windows from -550ms in
50ms time steps:

In:  -0.55 + 5 * 0.05
Out: -0.30000000000000004

In : -0.55 + 5 * 0.05 >= -0.3
Out: False

In : np.isclose(-0.55 + 5 * 0.05, -0.3)
Out: True