decim keyword for epochs fails if filter is None
(I guess I'm partially responsible for this myself, as I helped get BV importing done when no filter is set ...)

If info["lowpass"] is None, which can happen e.g. if importing underspecified BV files, and epoching uses the decim keyword, an error is thrown b/c ```new_sfreq < 2.5 * lowpass``` in line 140 fails if lowpass is None.
This is a simple workaround.

Might be worth considering if low/highpass = None should actually be allowed, or if import filters (specifically for BV) should default to something like 0 (or 1/length) and srate instead.