Resampling without losing events
Resampling the stim channel is a tricky business. If you are not
careful, you might lose events.

Consider the following stim data:

      y = [0, 0, 1, 0, 0, 0, 0, 1, 0, 0]

If we naively downsample by a factor of 2, this will become:

     y = [0, 1, 0, 0, 0]

whoops, event lost. The recommended approach is to find events on the
original data before resampling.

This PR adds an extra 'events' parameter to the `resample` method to resample the
data and event onsets jointly.

In addition some checks were added to warn the user if resampling might
cause problems in the stim channel.