Add outlines='skirt' option to plot_topomap
In some dense EEG montages, electrodes are at the side / underside of the
head. When maping a topomap plot, it can help interpretation to place
these electrodes (and corresponding interpolated data) outside the head
circle. Setting outline='skirt' allows you to do this.

EEGLAB has a similar feature, which is demonstrated here:
http://sccn.ucsd.edu/eeglab/allfunctions/topoplot.html