Add warnings for EEG electrodes not in montage
When a montage is applied, it is assumed that all the EEG channels in
the data have position information specified in the montage file. Raise
a warning if this is not the case.

Furthermore, the warning raised in `_auto_topomap_coords` when multiple
channels have the same position, has been extended to list the offending
channels.