Automatic info['nchan'] and info['ch_names']
The info object has two redundant fields: `nchan` and `ch_names`. They are
there as convenience fields. However, whenever the `chs` list is
updated, these fields need to be manually updated as well.

This PR makes these fields behave more like properties.

It does so by making `Info` a subclass of `collections.MutableMapping`, which allows
it to redefine `__setitem__` and `__getitem__` while retaining full
compatibility with the default Python dict.

The `nchan` field just maps to `len(info['chs'])`.

The `ch_names` field is a bit more tricky. From the outside, it behaves
as a mapping to `[ch['ch_name'] for ch in info['chs']]`. However, in order
not to generate a new list every time the field is accessed, the field
is an instance of `_ChannelNameList`.

The `_ChannelNameList` class is a subclass of `collections.Sequence`, thus
implementing a list that is read-only, but otherwise fully compatible
with a normal Python list. It overwrites the `__getitem(self, index)__` method to map
to `info['chs'][index]['ch_name']` on the fly.

The rest of the code is updated to no longer set the `nchan` and
`ch_names` fields of Info objects.