spatial colors z order by auc
For spatial_colors Butterfly plots, zorder matters. Currently, channel zorder is basically random - often, but not always, the nice red lower right channels are plotted last, and thus appear on top, giving the beautiful purple rainbow.

However, often, this means if you have two overlapping effects, the stronger one will occlude the weaker. This makes it so that the channels with the least activity are plotted on top.