Added get_params() method to mne.decoding
Added get_params() method to classes in CSP.py, time_frequency.py, and transformer.py. I did not add this method to Vectorizer class in transformer.py. The additions are untested.

With the get_params() method, it should now be possible to use scikit-learn's grid search on pipelines that employ MNE-Python functions.