[MRG+2] New DICS implementation (#5066)
* New DICS implementation

Taking @britta-wstnr 's implementation in #4430 as basis, we implemented
a new make_dics function. Notable differences with the original is that
it no longer requires a noise-CSD matrix (unit noise is assumed instead,
we can in a later PR tackle noise normalization again if need be), and
pick_ori='max-power' is now supported.

Also note that we made an enhancement to _lcmv._reg_pinv, namely the
rcond='auto' mode, to match the original MATLAB implementation. This
mode is necessary to make DICS behave properly and may be of some use to
LCMV beamformers as well.

In tutorials/plot_dics.py the first part of a DICS connectivity analysis
tutorial is provided. This part covers using DICS for finding two
simulated sources. In the upcoming PR's for connectivity analysis this
tutorial will be extended with one-to-all connectivity and finally
all-to-all connectivity analysis.

Applying a DICS beamformer to raw time-courses does have some practical
purpose, but it will not give you spectral source power over time (only
dics_source_power and tf_dics give you that). I'm afraid the current
plot_dics_beamformer.py example gives users the wrong idea on how to use
DICS. We can later add a proper example after deliberation with Jan
Kujala.

* Choose and document DICS defaults.

* PEP8

* Fix plot_tf_dics example

* Make tf_dics less verbose. Tweak plot_tf_dics example a bit.

* Rename "mode" parameter of make_dics to "inversion"

* Address comments

* Make tf_dics parameter list more similar to the original

* DOC: Comments

* Tweak docstrings of make_dics and apply_dics_csd

* Small nitpicks

* Remove unused import

* Update what's new