Update Travis Badge
The badge color does reflect the status of the master branch. However when clicking right now the badge points to the last run (which can be master or failing PR) Maybe pointing it to `mne-python/branches` is a better option. 

The other option, is to make `https://travis-ci.org/mne-tools/mne-python/` point always to master. (But I didn't find where to do so)