[MRG] LSLClient (#6120)
* working on the realtime client

* new base class for rtclient

* updates to the base client

* add import

* first pass at lsl in mne

* UPDATE lsl client, base_client, FIX epochs

* decouple the refactor from the LSLClient implementation

this reverts the attempt to both add an LSLClient and create a BaseClient class. For this upcoming PR, the focus will be adding support to LSL. subsequent PRs can be done to refactor the realtime module

* update to lsl client

at the moment, I am able to connect to a stream.
I need to test whether I am able to read the data from the stream.
opened an issue on pylsl to inquire whether i can both send and receive on same machine for local dev purpose (https://github.com/labstreaminglayer/liblsl-Python/issues/10)

once I can confirm the data is coming in properly, i will add some tests

* Update lsl_client.py

mistakenly tried to create inlets twice. this is the mvp for the LSL

* code cleanup and the beginning of tests for the lslclient

* Update test_lsl_client.py

* starting to remember the pythonic way of coding

been living in the world of javascript. feeling like a wayward youth

* simplify time

* Update lsl_client.py

* address some comment in the pr

change the pattern of checking for the import following common pattern seen in mne

* [FIX] _testing.py: copy pasted but never corrected

* Update test_lsl_client.py

forgot that I changed the class signature. If this doesn't work, I will implement multiprocessing to create the LSL Stream for the test

* testing is very useful

just caught the fact that I didn't change the base class to take the private connect function

* FIX arg in _connect; Implement thread to test

Fix a typo in the LSLClient._connect which led to downstream errors.

* address styling

* replace thread with process for clean exit

* linting complete

* FIX: Add to two other req lists

* FIX: Not on PyPi for Linux