ENH: add convenience function to select sources for simulation
To simulate cortical activity, the user must select a region of the
cortex to activate. This can be done with grow_labels, but requires
knowing the absolute index of the vertices.

This commit adds a new function select_sources which allows a user to
create a region based on a pre-existing label object.