MRG: add convenience function to select sources for simulation (#6193)
* ENH: add convenience function to select sources for simulation

To simulate cortical activity, the user must select a region of the
cortex to activate. This can be done with grow_labels, but requires
knowing the absolute index of the vertices.

This commit adds a new function select_sources which allows a user to
create a region based on a pre-existing label object.

* FIX: check parameter values using utility function

The value of the parameters of select_source can be checked using
_check_option to provide uniform error messages.

* REFACTOR: Move source selection function to label module

The select_sources function can be used in many situation, not only for
simulation.

This commit moves the select_sources function to the label module and
adds to the mne name space. The tests are also moved.

* MINOR: remove unused import

* TEST: Simplify tests to run in a few seconds

* MINOR: Replace np.isin by list comprehension

Version 1.11 of numpy does not support isin. This commit replaces it
by an equivalent list comprehension.

* MINOR: Add random state as a parameter to select_sources

* MINOR: Remove select_sources from the mne namespace and fix API ref

* Update whats_new.rst