FIX: fix bug where the length of the stc did not match duration
FIX: fix the duration property of SourceSimulator to return a float

When the duration is not provided to the SourceSimulator, it is computed
from the events. Instead of a duration in seconds (a float), it current
returns an integer.

This commit removes the cast to int to always return a float.

DOC: update whats_new.rst with SourceSimulator addition

DOC: add a short example on the SourceSimulator

This commit adds a short example that simulates source estimates and
raw data using the new SourceSimulator class.

MINOR: make code quality improvements suggested in PR #6212

ENH: change SourceSimulator API to use stop instead of n_samples