ENH: Include more realistic simulations to example for simulating raw data using subject anatomy (#6373)
* ENH: Include more realistic simulations to example for simulating raw data using subject anatomy

This commit changes the simulation of source time courses in the example for
simulating raw data using subject anatomy. Initially, sine wave was used for
source time courses and it was the same for each of the 4 simulation condition.
This commit adds a new function which enables to generate source activity for
evoked responses, parametrized by latency and duration. Stcs are simulated for
a single condition at the time.

The changes included in this commit are proposed in the discussion
of PR #6256 and #6261.

* FIX: Minor PEP8 issues.

This commit fixes minor PEP 8 issues.

* cosmits

* make it faster + reduce source amplitude to have more okish dSPM values