Fix ica.plot_properties with bad data (#6976)
* Fix ica.plot_properties with bad data

When plotting ICA component properties, one of the things that gets
plotted is the component PSD. This uses the helper function
`_convert_psds` that performs a check on the presense of zeros in the
PSD. When zeros are detected, a warning is printed. To construct the
warning message string, it is assumed that the first dimension of the `psds`
array contains channels. However, in the case of `ica.plot_properties`,
the first dimensions contains epochs. Hence, the construction of the
warning message string fails with an index out of bounds error.

This PR adds a flag to the `_convert_psds` helper function to inform it
that the first dimension of the `psds` array contains epochs rather than
channels.

* revert erroneous change