BF: Allow reading of FIF files with incorrect date
Currently, FIF files that have an invalid birthday or measurement date
set cannot be read, as an exception is raised (see #7492). This commit
catches the exception, skips the entry (effectively adding it as `None`
to the `info` dict), and emits a warning. This way, users can at least
load the data, even though the metadata may then be partially
incomplete.

Fixes GH-7492.