DOC: Make compute_raw_covariance easier to find
A short support request from @virvw made it obvious to me that
`mne.compute_raw_covariance` is currently not so easy to discover
as it should be; users were trying to use `mne.compute_covariance`
for empty-room data (and, consequently, running into trouble),
without realizing that `compute_raw_covariance` was readily
available. This commit adds a clearly-visible note to the docstring
to help guide users better.