[fNIRS]feat: add nosatflags_wlX files support
When the probes get saturated (i.e. when the value they return is "too
high"), NIRStar replaces the value with 'NaN' in the standard .wlX file.
The true value is kept in a .nosatflags_wlX file, which is a copy of a
.wlX file with the true probed values.

It is unclear how NIRstar decides that a probe got satured, so we added
a flag to `read_raw_nirx()` and `RawNIRX.__init__` to let the user chose
which file to use. Providing two *.wl1 or two *.wl2 files still raises
an error, as we check for the existence and unicity of the corresponding
*.nosatflags_wlX file.

Note that the directory structure check doesn't yet discriminate between
*.wlX and *.nosatflags_wlX (as it checks for any file ending with 'wlX').
Changing the check could be done in another PR to make sure the user did
not provide only *.nosatflags_wlX files and maybe warn them.