Better default legend in plot_compare_evokeds()
When passing a list of Evoked's to `viz.plot_compare_evokeds()`,
previously we would assign them sequential numbers (1 … N) as
labels, which is … not necessarily too helpful.

But since by default we're also storing useful information about the
condition(s) making up the Evoked in `Evoked.comment`, I thought
we could make use of this to create more useful labels.