Plot Evoked via joint_plot
Do not plot time traces and topomaps separately anymore. The joint plots
select the GFP peaks for plotting the topomaps, which is a much more
useful visualization than what we currently have in `master`.

Leave the old code (commented-out) so we can easily repurpose it later
on should we feel like it.