Fix changing line color in ICA.plot_overlay (#8341)
`ICA.plot_overlay()` didn't seem to work on my system with
latest Matplotlib anymore (macOS, both the default and Qt5Agg
backends). I would only see black traces. Debugging showed
that changig the colors of lines in a figure that has
been `show()`n was somehow part of the problem. Now we only
`show()` once all plotting operations are done, and this now
works with the default macOS and the Qt5Agg backends for me.