ENH: Allow automated compilation of metadata
This adds mne.epochs.make_metadata(), which allows for the automated
construction of metadata for use in Epochs. The users passes an
events array, an event name -> event ID mapping, a time period and
sampling frequency. This function will then iterate over all events
specified in the mapping, determine which events fall into the specified
time period (relative to the currently time-locked event), and will
add them to the metadata.