ENH: Slider triggers callback during dragging
The slider in the reports in `main` only triggers an update
callback when the dragging action has ended. This PR makes
it such that the callback is emitted during during slider movement,
so one doesn't have to release the mouse button to see the
results.

Also enables animations for clicking action.