Add qdarkstyle, darkdetect to environment.yml [circle full] (#9357)
I've added a conda-forge package for darkdetect, so it's now trivial to
include the packages we need for dark theme selection & automated
themesupport in our official conda environment – should be done anyway
so our CIs "see" this. Would be cool to get this merged before we make
the official 0.23 announcement.