MRG, ENH: Add NIRSport support (#9348)
* [fNIRS]feat: add nosatflags_wlX files support

When the probes get saturated (i.e. when the value they return is "too
high"), NIRStar replaces the value with 'NaN' in the standard .wlX file.
The true value is kept in a .nosatflags_wlX file, which is a copy of a
.wlX file with the true probed values.

It is unclear how NIRstar decides that a probe got satured, so we added
a flag to `read_raw_nirx()` and `RawNIRX.__init__` to let the user chose
which file to use. Providing two *.wl1 or two *.wl2 files still raises
an error, as we check for the existence and unicity of the corresponding
*.nosatflags_wlX file.

Note that the directory structure check doesn't yet discriminate between
*.wlX and *.nosatflags_wlX (as it checks for any file ending with 'wlX').
Changing the check could be done in another PR to make sure the user did
not provide only *.nosatflags_wlX files and maybe warn them.

* [fNIRS]feat: add nosatflags tests for NIRSport1

* [fNIRS]feat: add 'NaN' annotations

Some devices register 'NaN' values when the probes saturate. Since 'NaN'
values can cause unexpected behaviour, we added the ability to annotate
them so that they can be filtered out during the process.

* [fNIRS]feat: add a test for data with annotated_nan

* [fNIRS] update release number and hash following the addition of NIRSport v1 testing data

* Add to docs

* Improve wording and errors

* Use real data for tests

* Test that nans are returned. FAILING

* More tests. Still strange behaviour

* Found cause of strange behaviour

* Flake

* Add annotation description to docs

* Apply suggestions from code review

Co-authored-by: Eric Larson <larson.eric.d@gmail.com>

* Spelling

* Fix verbose

* Further suggestions from review

* Use new test file

* Flake

* Fix bug in logic

* Flake

* Doc

* doc

* Update utils.py

* FIX: Two-pass approach if needed

* FIX: Missed some

* ENH: Channel-specific

* FIX: Paired

* FIX: No need to ignore

* DOC: latest

Co-authored-by: David JULIEN <david.julien@ifsttar.fr>
Co-authored-by: Romain Derollepot <romain.derollepot@univ-eiffel.fr>
Co-authored-by: Eric Larson <larson.eric.d@gmail.com>