Add psd_signal_duration to Report.add_epochs()
The new parameter allows users to specify which duration of the signal
to use for PSD calculation. We will then pick N equally-spaced epochs
such that we cover this duration (or slightly more, due to the fixed
duration of invidiual epochs).

Still WIP because I haven't really tested it with "real" data yet; but
with `sample`, it can already speed up things significantly if I only
use 30 seconds of data; the resulting PSD figure looks pretty much
unchanged. So this could be a huge gain with only minor drawbacks.

The existing defaults from `main` are not changed; if the parameter
is `None`, we continue defaulting to using all epochs for PSD
calculation.