MRG: Make psd param of Report.add_epochs() accept floats as durations (#10119)
* Add psd_signal_duration to Report.add_epochs()

The new parameter allows users to specify which duration of the signal
to use for PSD calculation. We will then pick N equally-spaced epochs
such that we cover this duration (or slightly more, due to the fixed
duration of invidiual epochs).

Still WIP because I haven't really tested it with "real" data yet; but
with `sample`, it can already speed up things significantly if I only
use 30 seconds of data; the resulting PSD figure looks pretty much
unchanged. So this could be a huge gain with only minor drawbacks.

The existing defaults from `main` are not changed; if the parameter
is `None`, we continue defaulting to using all epochs for PSD
calculation.

* Better comments

* Fix exception message

* Avoid unnecessary copy if we're using all epochs

* Better logic

* Add test case

* Add changelog entry

* Use floor division

Co-authored-by: Alexandre Gramfort <alexandre.gramfort@m4x.org>

* Rename num_of_epochs_required -> n_epochs_required

* Extract into own method, handle edge case, add caption

* Improve tutorial

* Fix docstring

* Drop new parameter, make existing `psd` param accept floats

* Update mne/report/report.py

Co-authored-by: Alexandre Gramfort <alexandre.gramfort@m4x.org>