Fix scaling issue in Report.add_trans(), remove head transparency
We observed that the figure created by `Report.add_trans()`
was cropped. I narrowed this down to the `_ndarray_to_fig()`
function, which uses `Figure.figimage()`, which seems to later
interfere with our attempts to scale / adjust resolution of the
figure.

My fix is to not use `Figure.figimage()` anymore, and instead
add an axes that spans across the entire figure, and use `imshow()`.

With these changes, things seem to be looking good again.

It could be that this issue was introduced through #10129, but I haven't
invested much time to actually verify this hypothesis.

I have also removed transparency of the head surface in the
alighment plots – this makes it easier to visually inspect the
sensor alignment.