WIP: Copy BEM surfaces by default (don't symlink)
Several (all?) Windows users attenting a workshop we hosted yesterday
were running into issues when working with BEM surfaces we had shared
with them as part of a FreeSurfer subjects directory.

The reason was that transferring the symlinks turned out to be
problematic, and this users ended up with invalid files. They had to
manually copy over the BEM surfaces from the `bem/flash` or
`bem/watershed` (?) directories to `bem/` and rename them.

Since our collaborators are working in highly heterogenous settings
(macOS, Linux, Windows), this is indeed a huge problem for data
sharing, and it's not obvious at all to a common user WHY things are
not working as expected.

I therefore here propose an API change without deprecation cycle,
changing the default value of the `make_*_bem()` functions' `copy`
parameter from `False` to `True`, to always create a copy by default
(and not a symlink).

I'd further like to propose we change our sample data accordingly,
getting rid of the symbolic links for the BEM surfaces there as well.

I'm happy to discuss this at today's office hour.

cc @agramfort