[FIX] Remove _check_tfr_params from Transformer constructor (#11004)
* Moved _check_tfr_param to transform method

Since the _check_tfr_param function changes some of the values it
checks, scikit-learn estimators must perform this check at runtime, not
during estimator instantiation. Failing to do so leads to an error.

* Added test for cloning estimator with non-default parameter

If the constructor modifies non-default parameters, scikit-learn will
throw an error during cloning.

* Removed _check_tfr from constructor

Checking and transforming in the constructor violates sklearn
best-practice. Since this check is already performed in the
`_compute_tfr` method, we can simply remove the check from the
transformer class.

* Moved options checking to constructor

Since options checking doesn't modify the value it makes sense to keep
it in the constructor.

* Flake

* Remove blank line

* Updated docs

Co-authored-by: Daniel Carlström Schad <daniel.c.schad@protonmail.com>