Add error message when conversion of EEG locs to head space fails
When calling `.set_montage()`, all digitized points are transformed into
head coordinate space if possible. This is not possible when there are
no fiducials and the original points are not in head space already.
MNE-Python is lenient when the coordinate space of the original points
is marked as "unknown", assuming they are in head space. However, when
the points are explicitly marked as being in some other space (e.g.
MRI), then they will not get converted without fiducials.

Since MNE-Python requires all EEG channel positions to be in head space,
we should raise an error when `.set_montage()` fails to do this.

Fixes #10946.