Update _brain.py to allow usage of dict in legends
Related to the issue #11288. 

In the MNE documentation for the add_volume_labels() method from mne.viz.get_brain_class(), it's stated that you pass a dictionary to the legend option, which is actually implemented on the code. However, line 2475 from the _brain.py file actually does a wrong validation since it is set to _validate_type(legend, (bool, None), 'legend'); therefore, it actually doesn't allow using the dict option. I set it to validate_type(legend, (bool, None, dict), 'legend') in order to make it work.