DOC: Improve docstrings of multitaper functions
Some private functions in the mne.time_frequency.multitaper module had
unclear docstrings with missing array shapes and typos.

These changes add the array shapes of input and output arrays for the
functions in the module. Additionally, the name of the return value of
the _csd_from_mt function is corrected from "PSD" to "CSD".