Events outside EEG recording should not break the code
Some EGI events can be outside (before or after) the EEG recording. E.g. for video-synchronization markers or "IEND" codes. In this case, the code  breaks with an "index out of bounds" error. Examples are shown in the forum:
https://mne.discourse.group/t/index-error-when-trying-to-read-data/2823
https://mne.discourse.group/t/index-error-while-reading-data/5535
I had a similar issue with an EEG with video segments terminating later then the EEG recording.

The edits here should make sure to skip events that are placed illegally, i.e. below 0 (see forum example) or after the EEG end (other forum example+my case). This edit fixed the problem for me and should not harm the usual function. However, out-of-bounds events will be skipped/lost.
As alternative, the length of the events array could be extended for out of bounds events, but this may cause other issues with synchronizing to the EEG data.. Hence this option seems preferable. A warning may be added to notify the user of why an event was skipped.