ENH: log info when reading EDF/EDF+ files
Avoid warnings which might break backwards compatibility a bit too much.
Just log info when reading an EDF file that cannot be interpredted as EDF+,
or when the file is an EDF+ file, just not strictly compliant to EDF+.

Note that subfields of the header record can be replaced by a single
character `X`, espcially in de-identified files. This is legit and we
need not log anything in that case.

The `startdate` and `starttime` fields, on the other hand, cannot be
replaced with something that is not a date, like `X`. For the record,
the consensus in that case is to set them to `01.01.85` `00.00.00`.

Co-authored-by: Clemens Brunner <clemens.brunner@gmail.com>