Allow silencing of "The unit for channel(s) ... has changed" warnings
In MNE-BIDS we commonly change the units during an ordinary
reading operation. This is expected behavior and should not yield a
warning.

However, trying to catch and filter the warning MNE emits is quite
cumbersome, esp. inside the MNE-BIDS test suite.

I therefore believe the best approach is to allow silencing of the
warnings in MNE-Python directly.

This PR adds a new parameter, `on_unit_change`, to the
`.set_channel_types()` methods. The default is to warn, but users can
now also choose to `'ignore'` this.