ENH, STY: Refactor read_eyelink_calibration and update test After some deliberation, I believe that it is better to always return a Calibration instance for a single eye. This means that, for a recording in binocular mode, 2 Calibration instances will be returned for each "calibration" run during the session. For example if a single calibration was collected in binocular mode, the new API is just: Calibrations[0] for the left eye calibration, and Calibrations[1] for the right eye calibration. The two Calibration instances can be linked because they have the same exact onset.
my reasoning is:
Conceptual: even in binocular mode, a SINGLE calibration is done for each eye (though simultaneously). The calibration info for each eye is returned.
API: Users can expect a consistent structure in the Calibration class. For example, calibrations[0]["avg_error"] will alwasy be a single float (where before, if binocular, it was a dict).
Code Design: This makes the routines of read_eyelink_calibration much simpler, as we don't need a lot of conditional logic that depends on the recording mode being binocular or monocular.