ENH: additional parameters to calibration.plot method
Display the avg_offset and max_offset at the top of canvas

if show_offsets is True, display offset for each point

add param inver_y_axis, to invert y_axis. Defaults to True because in most monitors, pixel origin zero is at the top left corner of the screen. Thus, greater y coordinates are actually at the lower end of the screen.