FIX: error in old-ubuntu CI where dtype is a list
This is a strange one that I havent wrapped my head around.

From the stacktrace it seems like numpy_array.dtype is a list only in the Ci of old-ubuntu
where numpy 1.20 is used, and ubuntu 18 is used.

From the code I dont see anywhere that I could have accidently caste the dtype instance to
a list, and further, no error like this occurs in any other test.

- I added an isinsance(array.dtype, np.dtype) just before the line to see if this fixes the issue
- and confirms that this indeed was the problem line.

- Unrelated but I also added show_name == True in the test_plot_calibration test to add coverage to those lines of code