Add custom reference based on dict
The current implementation is based on a dictionary mapping a list to each channel. The list contains the channels to average and take as reference for the respective channel.
I have also implemented an assertion to make sure that the channels in the dict correspond to the channels in the instance.
We might want to implement something about bad channels. If bad channels are present in the mapping, they should probably be skipped