Refactor TimeChange, TimeBrowse, ChannelBrowse events.
Mouse and keypresses no longer directly update the plot, but instead
generate the appropriate UI events. Figures always listen and respond to
their own UI events, which is where the updating of the figure happens.
This way, there is a single place in the code where checking of corner
cases, etc. happens.

In addition, the matplotlib xdata timeline is different from actual
times in the Raw object. UI events should always be meaningful outside
of the figure. Hence, TimeChange and TimeBrowse should give the selected
times in terms of the original Raw object. Various translations
therefore need to happen.