Add a mne.io.BaseRAW.duration property
It represents the duration in seconds of the acquisition.

Centralize duration calculation for the textual (__repr__) and html
(_repr_html_, used by e.g. Jupyter) display of mne.io.BaseRAW instances
using the new property.

Extract the method to get the human string from the duration property
for use by the html representation, so it's easier to test later.