Fix mne.io.BaserRAW duration calculation
The duration calculation, used in `BaseRAW._html_repr_()` and
`BaseRAW.__repr__()`, was taking the timestamp of the last sample as the
duration of the acquisition, but was not accounting for the length of
the last sample.

Also, added tests for the refactored `BaseRAW.duration` property and
`BaseRAW._get_duration()` method, and used a sfreq value that revealed
the discrepancy in the duration calculation in the `BaseRAW.__repr__()`
method.

Finally, simplified the duration string calculation for the html display
by rounding up all the duration seconds, not just the remainder after
hour and minute calculations, thereby avoiding "00:01:60" calculations,
which should have been "00:02:00", when there are fractions of a second
remaining.

Fixes: #12954